after Authorization

+++ backend/routes/posts.js
@@ -134,9 +134,9 @@ router.post('', checkAuth, multer({storage: storage}).single("image") , (req, re
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
-
   post.save().then(createdPost => {
     res.status(201).json({
       message: 'Post added Sucessfully!',
@@ -158,11 +158,15 @@ router.post('', checkAuth, multer({storage: storage}).single("image") , (req, re
       _id : req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-    console.log(post);
-    Post.updateOne({_id: req.params.id}, post).then( result => {
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then( result => {
+     if( result.nModified > 0 ){
       res.status(200).json({message: "Update successfull!!" });
+     } else {
+      res.status(401).json({message: "Not Autherized" });
+     }
     });
   });
 
@@ -201,11 +205,12 @@ postQuery
   });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({
-      message: 'Post Successfully Deleted!'
-    });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if( result.n > 0 ){
+      res.status(200).json({message: "Deletion successfull!!" });
+     } else {
+      res.status(401).json({message: "Not Autherized" });
+     }
   });
 });
